Move mount modal to Modal file
