+++ library/src/scripts/modal/Modal.tsx
@@ -5,7 +5,7 @@
  */
 
 import ReactDOM from "react-dom";
-import React from "react";
+import React, { ReactElement } from "react";
 import classNames from "classnames";
 import ModalSizes from "@library/modal/ModalSizes";
 import { uniqueIDFromPrefix } from "@library/utility/idUtils";
@@ -49,6 +49,31 @@ interface IState {
 
 export const MODAL_CONTAINER_ID = "modals";
 export const PAGE_CONTAINER_ID = "page";
+
+/**
+ * Mount a modal with ReactDOM. This is only needed at the top level context.
+ *
+ * If you are already in a react context, just use `<Modal />`.
+ * Note: Using this will clear any other modals mounted with this component.
+ *
+ * @param element The <Modal /> element to render.
+ */
+export function mountModal(element: ReactElement<any>) {
+    // Ensure we have our modal container.
+    let modals = document.getElementById(MODAL_CONTAINER_ID);
+    if (!modals) {
+        modals = document.createElement("div");
+        modals.id = MODAL_CONTAINER_ID;
+        document.body.appendChild(modals);
+    } else {
+        ReactDOM.unmountComponentAtNode(modals);
+    }
+
+    ReactDOM.render(
+        element,
+        modals, // Who cares where we go. This is a portal anyways.
+    );
+}
 
 /**
  * An accessible Modal component.
